Javascript, HTML and CSS updates
Added new versions of jQuery and also a minified version for the same
Wrote some test code to update JSON data in a dropdown. This code will
be used in the datecoder form to update the drop down fields with new
options.
Created two gists for the generic JSON with and without AJAX
https://gist.github.com/fahdi/7553465 and
https://gist.github.com/fahdi/7553327
Added divider code for the dropdowns.

Note: Next version will contain the new code in datecoder.js file which
will use JSON data to update the respective format drop down fields on
change of an element dropdown. Excited!